Add replacement `ldapquery::search` function
Sourcing ldap server configuration options from puppet.conf was
conflating their original purpose, and a future release of Puppet may
even remove these options.

It's still desirable to be able to set defaults for the function from a
file, but a dedicated yaml file is far more flexible than an ini file.

In this commit, a new `ldapquery::search` function is added with a new
implementation. The old version is kept, but marked as DEPRECATED.